requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.7
django-on-heroku==1.1.2
gunicorn==20.0.4
html-table-parser-python3==0.1.5
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
